requirements.txt
agno==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.0.0
executing==2.2.0
feedparser==6.0.11
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-genai==1.11.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.2
lxml_html_clean==0.4.2
Markdown==3.8
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.2.4
openai==1.71.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
primp==0.14.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
rsa==4.9.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tldextract==5.2.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
